reorganize slack client into smaller files for each function group
